feat: coreEval to update priceFeeds, auction, and vaultManager
The primary goal is to stop using the original priceFeeds and
scaledPriceAuthorities that are retaining old quotePayments. (see

In order to do so, we also replace the auctioneer, and upgrade the
vaultManager, since they have no mechanism for dynamically finding out
about new priceAuthorities.

There are bootstrap tests and a3p-integration/ tests here to
demonstrate that after the upgrades and replacements, all these
well-known vats are in touch with one another and liquidation still
works.